requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
python-decouple==3.8
requests==2.28.2
ruff==0.7.2
soupsieve==2.6
urllib3==1.26.20
yarl==1.12.1
